+++ marketuca/src/modules/dashboard/components/Footer.jsx
@@ -3,7 +3,7 @@ import { motion } from "framer-motion";
 const Footer = () => {
   return (
     <motion.footer
-      className="py-8 bg-gray-100 relative"
+      className="py-8 bg-gray-100"
       initial={{ opacity: 0, y: 20 }}
       whileInView={{ opacity: 1, y: 0 }}
       viewport={{ once: true, amount: 0.3 }}
