+++ marketuca/src/modules/dashboard/components/Navbar.jsx
@@ -1,20 +1,11 @@
 import { useState } from "react";
-import {
-    Search,
-    ShoppingCart,
-    MessageSquare,
-    Heart,
-    User,
-    ShoppingBag,
-    LogOut,
-    BookUp, ClockAlert
-} from "lucide-react";
+import { Search, ShoppingCart, MessageSquare, Heart, User, ShoppingBag, LogOut } from "lucide-react";
 import { motion, AnimatePresence } from "framer-motion";
-import { Button } from "../ui/button";
-import { Input } from "../ui/input";
+import { Button } from "./ui/button";
+import { Input } from "./ui/input";
 import { Link } from "react-router-dom";
 
-const Navbar = ({ searchQuery, setSearchQuery, cartCount, isAdmin}) => {
+const Navbar = ({ searchQuery, setSearchQuery, cartCount }) => {
     const [isMenuOpen, setIsMenuOpen] = useState(false);
 
     const buttonVariants = {
@@ -71,28 +62,25 @@ const Navbar = ({ searchQuery, setSearchQuery, cartCount, isAdmin}) => {
             initial={{ y: -100 }}
             animate={{ y: 0 }}
             transition={{ duration: 0.5, ease: "easeOut" }}
-            className="sticky top-0 z-50 bg-white border-white-b shadow-sm rounded-md"
+            className="sticky top-0 z-50 bg-white border-b shadow-sm"
         >
             <div className="container flex items-center justify-between h-16 px-4 mx-auto">
 
-                <Link to={{ pathname: "/dashboard" }}>
-                    <motion.div
-                        className="flex items-center gap-2 select-none"
-                        whileHover={{scale: 1.05}}
-                        transition={{duration: 0.3}}
-                    >
-                        <motion.img
-                            variants={owlVariants}
-                            animate="animate"
-                            className="h-8 w-8 sm:h-10 sm:w-10"
-                            src="/buho.png"
-                            alt="MarketPlace UCA Logo"
-                        />
-                        { /*<img src="/buho.png" alt="Logo búho" className="w-6 h-6" />*/}
-                        <span
-                            className="text-xl font-bold text-[#0056b3]">MarketPlace UCA</span>
-                    </motion.div>
-                </Link>
+                <motion.div
+                    className="flex items-center gap-2"
+                    whileHover={{ scale: 1.05 }}
+                    transition={{ duration: 0.3 }}
+                >
+                    <motion.img
+                        variants={owlVariants}
+                        animate="animate"
+                        className="h-8 w-8 sm:h-10 sm:w-10"
+                        src="/buho.png"
+                        alt="MarketPlace UCA Logo"
+                    />
+                    { /*<img src="/buho.png" alt="Logo búho" className="w-6 h-6" />*/}
+                    <span className="text-xl font-bold text-[#0056b3]">MarketPlace UCA</span>
+                </motion.div>
                 <div className="flex items-center max-w-md w-full relative">
                     <Input
                         type="text"
@@ -105,7 +93,13 @@ const Navbar = ({ searchQuery, setSearchQuery, cartCount, isAdmin}) => {
                     <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 w-4 h-4 text-gray-400" />
                 </div>
                 <div className="flex items-center gap-3">
-
+                    <motion.div whileHover="hover" whileTap="tap" variants={buttonVariants}>
+                        <Button variant="ghost" size="icon" className="rounded-full">
+                            <Link to="/">
+                                <MessageSquare className="w-5 h-5 text-[#0056b3]" />
+                            </Link>
+                        </Button>
+                    </motion.div>
                     <motion.div whileHover="hover" whileTap="tap" variants={buttonVariants}>
                         <Button variant="ghost" size="icon" className="rounded-full">
                             <Link to="/favorites">
@@ -153,11 +147,11 @@ const Navbar = ({ searchQuery, setSearchQuery, cartCount, isAdmin}) => {
                                             className="flex items-center gap-2 px-4 py-2 text-gray-700 cursor-pointer"
                                             onClick={() => {
                                                 setIsMenuOpen(false);
-                                                window.location.href = "/profile"; // Recarga la página al redirigir
+                                                window.location.href = "/my-items"; // Recarga la página al redirigir
                                             }}
                                         >
                                             <ShoppingBag className="w-4 h-4" />
-                                            <span>Mi perfil</span>
+                                            <span>Mis artículos</span>
                                         </motion.div>
                                         <motion.div
                                             variants={menuItemVariants}
@@ -171,41 +165,12 @@ const Navbar = ({ searchQuery, setSearchQuery, cartCount, isAdmin}) => {
                                             <Heart className="w-4 h-4" />
                                             <span>Mis favoritos</span>
                                         </motion.div>
-                                        {isAdmin && (
-                                            <motion.div
-                                                variants={menuItemVariants}
-                                                whileHover="hover"
-                                                className="flex items-center gap-2 px-4 py-2 text-gray-700 cursor-pointer"
-                                                onClick={() => {
-                                                    setIsMenuOpen(false);
-                                                    window.location.href = "/waitlist";
-                                                }}
-                                            >
-                                                <ClockAlert className="w-4 h-4" />
-                                                <span>Lista de espera</span>
-                                            </motion.div>
-                                        )}
-                                        {isAdmin && (
-                                            <motion.div
-                                                variants={menuItemVariants}
-                                                whileHover="hover"
-                                                className="flex items-center gap-2 px-4 py-2 text-gray-700 cursor-pointer"
-                                                onClick={() => {
-                                                    setIsMenuOpen(false);
-                                                    window.location.href = "/adminman";
-                                                }}
-                                            >
-                                                <BookUp className="w-4 h-4" />
-                                                <span>Manejo de administrador</span>
-                                            </motion.div>
-                                        )}
                                         <motion.div
                                             variants={menuItemVariants}
                                             whileHover="hover"
                                             className="flex items-center gap-2 px-4 py-2 text-gray-700 cursor-pointer border-t border-gray-200"
                                             onClick={handleLogout}
                                         >
-
                                             <LogOut className="w-4 h-4" />
                                             <span>Cerrar sesión</span>
                                         </motion.div>
