+++ marketuca/src/modules/dashboard/components/ProductsSection.jsx
@@ -2,16 +2,7 @@ import { motion, useInView } from "framer-motion";
 import { useRef, useEffect } from "react";
 import { Heart } from "lucide-react";
 
-const ProductsSection = ({
-                             products,
-                             loading,
-                             hasMore,
-                             loadMoreProducts,
-                             activeCategory,
-                             onProductClick,    // <-- Nombre estándar del prop
-                             favorites,
-                             toggleFavorite,
-                         }) => {
+const ProductsSection = ({ products, loading, hasMore, loadMoreProducts, activeCategory, handleProductClick, favorites, toggleFavorite }) => {
     const productVariants = {
         hidden: { opacity: 0, y: 20 },
         visible: {
@@ -60,16 +51,10 @@ const ProductsSection = ({
                         whileInView="visible"
                         viewport={{ once: true, amount: 0.3 }}
                         whileHover={{ scale: 1.03, boxShadow: "0 10px 20px rgba(0, 86, 179, 0.15)", borderColor: "#339CFF" }}
-                        onClick={() => onProductClick(product)}
+                        onClick={() => handleProductClick(product)}
                     >
                         <div className="relative overflow-hidden aspect-square">
-                            <motion.img
-                                src={product.image}
-                                alt={product.title}
-                                className="object-cover w-full h-full"
-                                whileHover={{ scale: 1.05 }}
-                                transition={{ duration: 0.3 }}
-                            />
+                            <motion.img src={product.image} alt={product.title} className="object-cover w-full h-full" whileHover={{ scale: 1.05 }} transition={{ duration: 0.3 }} />
                             <motion.button
                                 className={`absolute top-2 right-2 rounded-full p-2 bg-white/80 backdrop-blur-sm ${favorites.includes(product.id) ? "text-red-500" : "text-gray-500"}`}
                                 onClick={(e) => {
@@ -89,14 +74,14 @@ const ProductsSection = ({
                         <div className="p-4">
                             <div className="flex items-start justify-between gap-2">
                                 <h3 className="font-semibold text-gray-800">{product.title}</h3>
-                                <p className="text-lg font-bold text-[#0056b3]">${product.price?.toFixed(2) ?? "0.00"}</p>
+                                <p className="text-lg font-bold text-[#0056b3]">${product.price.toFixed(2)}</p>
                             </div>
                             <div className="flex items-center mt-2 text-sm text-gray-500">
                                 <span>{product.condition}</span>
                                 <span className="mx-2">•</span>
                                 <div className="flex items-center">
                                     <span className="text-yellow-500">★</span>
-                                    <span className="ml-1">{product.rating ?? "4.5"}</span>
+                                    <span className="ml-1">{product.rating}</span>
                                 </div>
                             </div>
                             <div className="flex items-center justify-between mt-4">
@@ -106,7 +91,7 @@ const ProductsSection = ({
                                         className="bg-gradient-to-r from-[#0056b3] to-[#339CFF] hover:from-[#339CFF] hover:to-[#0056b3] text-white px-3 py-1 rounded"
                                         onClick={(e) => {
                                             e.stopPropagation();
-                                            onProductClick(product);
+                                            handleProductClick(product);
                                         }}
                                     >
                                         Ver detalles
