+++ marketuca/src/modules/dashboard/page/dashboard.jsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from "react";
 import Particles from "react-tsparticles";
-import Navbar from "../../utils/navbar/Navbar.jsx";
+import Navbar from "../components/Navbar";
 import HeroSection from "../components/HeroSection";
 import CategoriesSection from "../components/CategoriesSection";
 import ProductsSection from "../components/ProductsSection";
-import Footer from "../../utils/footer/Footer.jsx";
+import Footer from "../components/Footer";
 import { useDashboard } from "../hooks/useDashboard";
 import { ProductDetail } from "../components/modal/product-detail";
 import { SellProductModal } from "../components/modal/sell-product-modal";
@@ -22,8 +22,6 @@ import {
   User,
   ShirtIcon,
 } from "lucide-react";
-import ParticlesDashboard
-  from "../../utils/ui/ParticlesDashboard.jsx";
 
 export default function Dashboard() {
   const {
@@ -84,14 +82,36 @@ export default function Dashboard() {
         `}
       </style>
 
-      <ParticlesDashboard/>
+      <Particles
+        id="tsparticles"
+        options={{
+          background: { color: { value: "transparent" } },
+          fpsLimit: 60,
+          particles: {
+            number: { value: 80, density: { enable: true, value_area: 1500 } },
+            color: { value: ["#007BFF", "#FFFFFF", "#00C4B4"] },
+            shape: { type: "circle" },
+            opacity: { value: 0.6, random: true, anim: { enable: true, speed: 1, opacity_min: 0.2, sync: false } },
+            size: { value: 5, random: true, anim: { enable: true, speed: 2, size_min: 0.5, sync: false } },
+            line_linked: { enable: false },
+            move: { enable: true, speed: 1.5, direction: "none", random: true, straight: false, out_mode: "out", bounce: false },
+          },
+          interactivity: {
+            detect_on: "canvas",
+            events: { onhover: { enable: false, mode: "repulse" }, onclick: { enable: false, mode: "push" }, resize: true },
+            modes: { repulse: { distance: 100, duration: 0.4 }, push: { particles_nb: 4 } },
+          },
+          retina_detect: true,
+        }}
+        className="absolute top-0 left-0 w-full h-full pointer-events-none"
+        style={{ zIndex: 0 }}
+      />
 
       <div className="relative min-h-screen">
         <Navbar
           cartCount={cart.reduce((total, item) => total + item.quantity, 0)}
           searchQuery={searchQuery}
           setSearchQuery={setSearchQuery}
-          isAdmin={true}
         />
         <HeroSection onSellClick={() => setIsSellModalOpen(true)} />
         <CategoriesSection
