+++ marketuca/src/modules/home/components/Hero.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { motion } from 'framer-motion';
 import Particles from 'react-tsparticles';
-import ParticlesBackground from "../../utils/ParticlesBackground.jsx";
 
 const Hero = () => {
   // Variantes para animaciones de entrada
@@ -17,7 +16,88 @@ const Hero = () => {
 
   return (
 <section className="relative pb-20">
-      <ParticlesBackground/>
+<Particles
+        id="tsparticles"
+        options={{
+          background: {
+            color: {
+              value: '#f0f4f8'
+            }
+          },
+          particles: {
+            number: {
+              value: 80,
+              density: {
+                enable: true,
+                value_area: 700
+              }
+            },
+            color: {
+              value: ['#007BFF', '#FFFFFF', '#00C4B4']
+            },
+            shape: {
+              type: 'circle'
+            },
+            opacity: {
+              value: 0.6,
+              random: true,
+              anim: {
+                enable: true,
+                speed: 1,
+                opacity_min: 0.2,
+                sync: false
+              }
+            },
+            size: {
+              value: 5,
+              random: true,
+              anim: {
+                enable: true,
+                speed: 2,
+                size_min: 0.5,
+                sync: false
+              }
+            },
+            line_linked: {
+              enable: false
+            },
+            move: {
+              enable: true,
+              speed: 1.5,
+              direction: 'none',
+              random: true,
+              straight: false,
+              out_mode: 'out',
+              bounce: false
+            }
+          },
+          interactivity: {
+            detect_on: 'canvas',
+            events: {
+              onhover: {
+                enable: true,
+                mode: 'repulse'
+              },
+              onclick: {
+                enable: true,
+                mode: 'push'
+              },
+              resize: true
+            },
+            modes: {
+              repulse: {
+                distance: 100,
+                duration: 0.4
+              },
+              push: {
+                particles_nb: 4
+              }
+            }
+          },
+          retina_detect: true
+        }}
+        className="absolute inset-0 z-0 pointer-events-none"
+      />
       <div className="relative z-10 w-full px-4 sm:px-6 md:px-20 lg:px-24 xl:px-32 flex flex-col md:flex-row items-center justify-between gap-4 sm:gap-6 md:gap-12 pt-32 sm:pt-36 md:pt-44">
         <div className="w-full md:w-1/2 text-center md:text-left mb-10 sm:mb-12 md:mb-16 lg:mb-20 2xl:mb-24">
           <motion.h1
