+++ marketuca/src/modules/home/page/home.jsx
@@ -1,16 +1,16 @@
 // Home.jsx
+import Navbar from '../components/Navbar';
 import Hero from '../components/Hero';
 import Resources from '../components/Resources';
 import FranjaAzul from '../components/FranjaAzul';
 import HowItWorks from '../components/HowItWorks';
 import Benefits from '../components/Benefits';
-import Footer from '../../utils/footer/Footer.jsx';
-import NavbarHome from "../../utils/navbar/NavbarHome.jsx";
+import Footer from '../components/Footer';
 
 const Home = () => {
   return (
     <div>
-      <NavbarHome />
+      <Navbar />
       <Hero />
       <FranjaAzul />
       <Resources />
