--- marketuca/src/modules/register/page/register.jsx
@@ -1,17 +1,0 @@
-import NavbarNoAuth from "../../utils/navbar/NavBarNoAuth.jsx";
-import Footer from "../../utils/footer/Footer.jsx";
-import RegisterForm from "../components/RegisterForm.jsx";
-
-const register = () => {
-
-
-    return(
-        <div>
-            <NavbarNoAuth/>
-            <RegisterForm/>
-            <Footer/>
-        </div>
-    );
-
-}
-export default register;
